requirements.txt
tensorflow==2.13.0
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.3
pandas==2.0.3
Pillow==10.0.0
